feat(public): 添加新图片资源并删除旧图片
refactor(components): 重构组件结构，将视图组件移至组件目录

style(HeaderComponent): 调整头部组件样式

chore(.gitignore): 更新.gitignore文件，添加public/对应名单.txt

refactor(App.vue): 启用RouterView并移除注释代码

refactor(main.ts): 注释掉Pinia的引入和使用

feat(ArticleDetailComponent): 新增文章详情组件

refactor(router): 添加WorkView路由并移除BookView路由

refactor(HomeView): 重构HomeView，使用新的组件结构

feat(AboutComponent): 新增关于我们组件

refactor(ContactComponent): 新增联系我们组件

feat(WorkView): 新增作品展示视图

refactor(NewsComponent): 新增新闻组件

refactor(CursorComponent): 新增光标组件

feat(ServerComponent): 新增服务组件

feat(ShowComponent): 新增作品展示组件
