+++ components.d.ts
@@ -8,14 +8,31 @@ export {}
 /* prettier-ignore */
 declare module 'vue' {
   export interface GlobalComponents {
+    AboutComponent: typeof import('./src/components/AboutComponent.vue')['default']
+    AboutComponents: typeof import('./src/components/AboutComponents.vue')['default']
+    AboutView: typeof import('./src/components/AboutView.vue')['default']
     ArticleDetail: typeof import('./src/components/ArticleDetail.vue')['default']
+    ArticleDetailComponent: typeof import('./src/components/ArticleDetailComponent.vue')['default']
+    ArticleDetailComponents: typeof import('./src/components/ArticleDetailComponents.vue')['default']
+    ContactComponent: typeof import('./src/components/ContactComponent.vue')['default']
+    ContactComponents: typeof import('./src/components/ContactComponents.vue')['default']
+    ContactView: typeof import('./src/components/ContactView.vue')['default']
+    copy: typeof import('./src/components/AboutComponents copy.vue')['default']
+    CursorComponent: typeof import('./src/components/CursorComponent.vue')['default']
     CursorComponents: typeof import('./src/components/CursorComponents.vue')['default']
     ElTimeline: typeof import('element-plus/es')['ElTimeline']
     ElTimelineItem: typeof import('element-plus/es')['ElTimelineItem']
     FooterComponent: typeof import('./src/components/FooterComponent.vue')['default']
     HeaderComponent: typeof import('./src/components/HeaderComponent.vue')['default']
+    NewsComponent: typeof import('./src/components/NewsComponent.vue')['default']
+    NewsComponents: typeof import('./src/components/NewsComponents.vue')['default']
+    NewsView: typeof import('./src/components/NewsView.vue')['default']
     PreloadComponents: typeof import('./src/components/PreloadComponents.vue')['default']
     RouterLink: typeof import('vue-router')['RouterLink']
     RouterView: typeof import('vue-router')['RouterView']
+    ServerComponent: typeof import('./src/components/ServerComponent.vue')['default']
+    ServerComponents: typeof import('./src/components/ServerComponents.vue')['default']
+    ServerView: typeof import('./src/components/ServerView.vue')['default']
+    ShowComponent: typeof import('./src/components/ShowComponent.vue')['default']
   }
 }
