+++ src/App.vue
@@ -4,7 +4,7 @@
   <div>
     <!-- <CursorComponents /> -->
     <HeaderComponent />
-    <!-- <HomeView /> -->
+    <RouterView />
     <!-- 将 footer 移到 app-container 内部 -->
     <!-- <FooterComponent /> -->
   </div>
