+++ src/router/index.ts
@@ -1,6 +1,6 @@
 import { createRouter, createWebHashHistory } from 'vue-router'
 import HomeView from '@/views/HomeView.vue'
-// import BookView from '@/views/BookView.vue'
+import WorkView from '@/views/WorkView.vue'
 // import NewsView from '@/views/NewsView.vue'
 // import AboutView from '@/views/AboutView.vue'
 // import ContactView from '@/views/ContactView.vue'
@@ -26,11 +26,11 @@ const router = createRouter({
     //   name: 'news',
     //   component: NewsView,
     // },
-    // {
-    //   path: '/book',
-    //   name: 'book',
-    //   component: BookView,
-    // },
+    {
+      path: '/work',
+      name: 'work',
+      component: WorkView,
+    },
     // {
     //   path: '/contact',
     //   name: 'contact',
