+++ src/views/HomeView.vue
@@ -1,18 +1,19 @@
 <template>
   <!-- 主容器，用于将内部内容居中显示 -->
   <div id="home-container">
-    <AboutView />
-    <BookView />
-    <ServerView />
-    <NewsView />
-    <ContactView />
+    <ShowComponent />
+    <NewsComponent />
+    <AboutComponent />
+    <!-- <BookView /> -->
+    <ServerComponent />
+    <ContactComponent />
   </div>
 </template>
 
 <script setup lang="ts">
-import AboutView from './AboutView.vue'
-import BookView from './BookView.vue'
-import ServerView from './ServerView.vue'
-import NewsView from './NewsView.vue'
-import ContactView from './ContactView.vue'
+import AboutComponent from '@/components/AboutComponent.vue'
+import ServerComponent from '@/components/ServerComponent.vue'
+import NewsComponent from '@/components/NewsComponent.vue'
+import ContactComponent from '@/components/ContactComponent.vue'
+import ShowComponent from '@/components/ShowComponent.vue'
 </script>
